Changes to util classes to add some flexibility to the location of the PresentatioonState object.
Use that flexibility to keep the presentation state in the module state. Ensure it's embedded in the request and get information from that (appRoot etc).

This should allow a module to set the skin name without affecting the rest of the ui.

Also chnage from referring to "client" to "module"