Determined that we couldn't run with the indexes split into public/user. Need a single index as aliases can cross from user to public.
A number of db methods were duplicated with indexer equivalents

Indexer was set up to skip aliases - it would follow the aliases as regular collections and try to index data twice.

Significant changes to access checking. Required much refactoring. Moved it down a couple of levels. It's now a constructor parameter for the indexer - which used to check only on searh.

Use access checker to check when retrieving collections.